moved jest tst files to test folder. running
