--- grpc_mockData/protos/helloworld.proto
@@ -1,15 +1,0 @@
-// now refactored to work with testServer.js in the newer version of grpc (@grpc/grpc-js)
-
-syntax= "proto3";
-
-service HelloWorldService {
-  rpc GreetMe (GreetRequest) returns (GreetReply) {}
-}
-
-message GreetRequest {
-  string name = 1;
-}
-
-message GreetReply {
-  string reply = 1;
-}
