--- grpc_mockData/protos/hw2_multi_service.proto
@@ -1,29 +1,0 @@
-syntax = "proto3";
-
-package helloworld;
-
-// The greeting service definition.
-service Greeter {
-  // Sends a greeting
-  rpc SayHello (HelloRequest) returns (HelloReply) {}
-  rpc SayHelloCS (stream HelloRequest) returns (HelloReply) {}
-  rpc SayHellos (HelloRequest) returns (stream HelloReply) {}
-  rpc SayHelloBidi (stream HelloRequest) returns (stream HelloReply) {}
-}
-
-service Butler {
-  // Sends a greeting
-  rpc SayHello (HelloRequest) returns (HelloReply) {}
-  rpc SayHelloCS (stream HelloRequest) returns (HelloReply) {}
-  rpc SayHellos (HelloRequest) returns (stream HelloReply) {}
-  rpc SayHelloBidi (stream HelloRequest) returns (stream HelloReply) {}
-}
-// The request message containing the user's name.
-message HelloRequest {
-  string name = 1;
-}
-
-// The response message containing the greetings
-message HelloReply {
-  string message = 1;
-}
