+++ test/__mocks__/electronMock.js
