+++ test/__tests__/protoParserTests.js
@@ -1,4 +1,4 @@
-import protoParser from '../main_process/protoParser';
+import protoParser from '../../main_process/protoParser';
 
 describe('testing protoParser', () => {
   const protoFile = `syntax = 'proto3';
